Check task attribute before use in sentry.add_tagging() (#37143)
* Check task attribute before use in add_tagging

* Refactor sentry and add tests

---------
